RAG/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain e dependências
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
langchain-core==0.1.10

# Banco de dados
sqlite-vss==0.1.2

# Processamento de dados
pandas==2.1.4
numpy==1.25.2
numexpr==2.8.7

# Utilitários
jellyfish==1.2.0
python-dotenv==1.0.0

# CORS e middleware
python-multipart==0.0.6

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 